% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotate.R
\name{add_annotate}
\alias{add_annotate}
\title{\code{add_annotate()} is used in combination with \code{whomapper()} and \code{bubblemapper()},
and adds optional text and shapes located at the centroid of listed countries and territories, with adjustment for CRS setting.}
\usage{
add_annotate(
  geom,
  x = NULL,
  y = NULL,
  xend = NULL,
  yend = NULL,
  xmin = NULL,
  ymin = NULL,
  xmax = NULL,
  ymax = NULL,
  linewidth = NULL,
  size = NULL,
  color = "black",
  label = NULL,
  include_coord = FALSE,
  default_crs = 4326
)
}
\arguments{
\item{size}{Marker size.}

\item{iso3}{A vector of ISO3 country codes.}

\item{shape}{Marker shape.}

\item{col}{Marker colour.}

\item{lab}{Marker label.}
}
\value{
A ggplot2 function.
}
\description{
\code{add_annotate()} is used in combination with \code{whomapper()} and \code{bubblemapper()},
and adds optional text and shapes located at the centroid of listed countries and territories, with adjustment for CRS setting.
}
\examples{
add_annotate("text", x = -50, y = 10, label = "Example", size = 4, color = "blue")
}
\author{
Takuya Yamanaka.
}
