% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marker.R
\name{add_marker}
\alias{add_marker}
\title{\code{add_marker()} is used in combination with \code{whomapper()} and adds a secondary
marker located at the centroid of listed countries and territories.}
\usage{
add_marker(
  iso3 = NA_character_,
  shape = 17,
  col = "red",
  size = 3,
  alpha = 1,
  lab = "",
  projection = "robin",
  offset = 10.8
)
}
\arguments{
\item{iso3}{A vector of ISO3 country codes.}

\item{shape}{Marker shape.}

\item{col}{Marker colour.}

\item{size}{Marker size.}

\item{lab}{Marker label.}
}
\value{
A ggplot2 function.
}
\description{
\code{add_marker()} is used in combination with \code{whomapper()} and adds a secondary
marker located at the centroid of listed countries and territories.
}
\examples{
brics <- data.frame(iso3=c('BRA','CHN','IND','RUS','ZAF'), var=1:5)
whomap(brics, legend.title='BRICS', legend.pos=c(0.14, 0.34)) +
   add_marker(c('BRA','RUS'), lab='Subnational\ndata')

}
\author{
Takuya Yamanaka, adapted from scripts of whomap developed by Philippe Glaziou.
}
