% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_builder.R
\name{common_disputed_border}
\alias{common_disputed_border}
\title{Common map layers}
\usage{
common_disputed_border(
  p,
  layers,
  map_title,
  disclaimer,
  legend_pos,
  line_col,
  line_width,
  water_col,
  china_color,
  korea_color,
  sudan_color,
  palestine_color,
  disclaim
)
}
\description{
Common map layers
}
