% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomapper.R
\name{whomapper}
\alias{whomapper}
\title{Choropleth world maps 2025 version
\code{whomapper()} prints a choropleth world map based on the latest WHO geoJSON files
It requires ggplot2, ggpattern, sf and here}
\source{
Modified from WHO GIS (https://gis-who.hub.arcgis.com/)
}
\usage{
whomapper(
  X = data.frame(iso3 = NA, var = NA),
  colours = NULL,
  projection = "eqearth",
  offset = NULL,
  low_col = "#BDD7E7",
  high_col = "#08519C",
  line_col = "black",
  line_width = 0.3,
  map_title = "",
  legend_title = "",
  water_col = "white",
  na_label = "No data",
  na_col = "white",
  disclaimer = FALSE,
  legend_pos = c(0.17, 0.42),
  zoom = "Global"
)
}
\arguments{
\item{X}{A dataframe with two columns: 'iso3' (character or factor WHO country codes) and 'var' (categorical variable to map).}

\item{colours}{A vector of colour values for each category in "var", excepting missing values.}

\item{low_col}{First value of a gradient of colours.}

\item{high_col}{Last value of a gradient of colours.}

\item{line_col}{Colour of country border lines.}

\item{map_title}{Map title.}

\item{legend_title}{Legend title.}

\item{water_col}{Colour of oceans and lakes.}

\item{na_label}{Legend lable for missing values.}

\item{na_col}{Colour of countries with missing values.}

\item{disclaimer}{A boolean, inserts a standard WHO disclaimer.}

\item{legend_pos}{A vector of two numbers, positions the legend.}
}
\value{
A ggplot2 plot.
}
\description{
with standard WHO ISO3 country codes.The categorical variable to be
mapped should be named "var" (see examples).
}
\examples{
whomapper(data.frame(iso3 = NA, var = NA))
}
\author{
Takuya Yamanaka, adapted from scripts of whomap developed by Philippe Glaziou.
}
